feat: add exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,9 @@ public class ProductService {
     // apenas leitura pois o metodo é get
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        var product = repository.findById(id).get();
+        var product = repository.findById(id).orElseThrow(
+            () -> new ResourceNotFoundException("Recurso não encontrado.")
+        );
         return new ProductDTO(product);
     }
 
@@ -44,17 +51,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        var entity = repository.getReferenceById(id);
+        try {
+            var entity = repository.getReferenceById(id);
 
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException exception) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException exception) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
